refactor: Refactor FAQ panel logic, enhance ListItem and HistoryList, fix Popover rendering and referral button loading state(OK-39661 OK-39660) (#7683)
* Refactor FAQ panel rendering logic in EarnHome

Updated the conditional rendering of the FAQ panel to better account for the presence of banners and screen size. Removed redundant code and improved clarity in the display logic.

* Support ReactNode for ListItem subtitle and enhance HistoryList

Updated ListItem to accept a subtitle of type string or ReactNode, enabling richer content. Enhanced the Staking HistoryList to display a shortened account address with an explorer link when filterType is 'rebate', improving user experience and flexibility.

* Fix Popover rendering and trigger handling for native

Wraps the floating panel in a platform check to prevent rendering on native platforms and adds a conditional check for renderTrigger in BasicPopover to avoid rendering an empty trigger. This improves compatibility and prevents errors on native environments.

* Fix loading state condition in referral code button

Changed the conditional logic to render the skeleton item only when both displayReferralCodeButton is undefined and isLoading is true, preventing premature rendering.